bin/requirements.txt
# Core dependencies
python-dotenv
mcp
fastmcp
PyYAML
pandas
pytest-asyncio
asyncio

# LLM Provider SDKs
google-generativeai
openai
anthropic
ollama 